Rename HttpRequest.Url to HttpRequest.RequestUrl
Otherwise gcc will break with the Url class, clang works.

refs #9447

+++ lib/remote/httpchunkedencoding.cpp
@@ -51,7 +51,7 @@ StreamReadStatus HttpChunkedEncoding::ReadChunkFromStream(const Stream::Ptr& str
 			scontext.MustRead = false;
 		}
 
-		if (scontext.Size < context.LengthIndicator) {
+		if (scontext.Size < (size_t)context.LengthIndicator) {
 			scontext.MustRead = true;
 			return StatusNeedData;
 		}
